realtime/collectors-secondary/requirements.txt
websockets>=12.0
aiohttp>=3.9
redis[hiredis]>=5.0
python-dotenv>=1.0
